More config work

+++ src/main/java/sourcecoded/core/configuration/VersionConfig.java
@@ -1,11 +1,10 @@
 package sourcecoded.core.configuration;
 
+import org.apache.commons.io.FileUtils;
+import sourcecoded.core.util.SourceLogger;
+
 import java.io.File;
 import java.io.IOException;
-
-import org.apache.commons.io.FileUtils;
-
-import sourcecoded.core.util.SourceLogger;
 
 /**
  * A configuration class extending SourceConfig
@@ -39,28 +38,23 @@ public class VersionConfig extends SourceConfig {
     }
 
     private static VersionConfig setupVersioning(VersionConfig config) throws IOException {
+        config.createProperty(CATEGORY_VERSION, "ConfigRevision", config.targetVersion);
+        config.saveConfig();
 
+        config.config.addCustomCategoryComment(CATEGORY_VERSION, "Used for Configuration Version Control. DO NOT CHANGE THIS.");
         String currentVersion = config.getString(CATEGORY_VERSION, "ConfigRevision");
 
         if (!currentVersion.equals(config.targetVersion)) {
-            configLogger.error(String.format("Configuration File for Mod: %s is out of date! (target %s, found %s). Trying to move the old configuration file to %s_old", config.modid, config.targetVersion, currentVersion, config.configFile.getName()));
+            configLogger.error(String.format("Configuration File for Mod: %s is out of date! (target %s, found %s). The old configuration has been moved to /configOld/%s_old", config.modid, config.targetVersion, currentVersion, config.configFile.getName()));
 
-            File target = new File(config.configFile.getAbsolutePath() + "_old");
+            File newDir = new File(config.configFile.getPath() + "/../../configOld");
+            File target = new File(newDir + "/" + config.configFile.getName() + "_old");
 
-            if (target.exists()) {
-                target.delete();
-            }
+            if (target.exists()) target.delete();
+            FileUtils.moveFile(config.configFile, target);
 
-            try {
-                FileUtils.moveFile(config.configFile, target);
-            } catch (IOException e) {
-                configLogger.error("Failed to move the old configuration file to %s" + target.getName());
-                return config;
-            }
-
-            config.createProperty(CATEGORY_VERSION, "ConfigRevision", config.targetVersion).set(config.targetVersion);
-            config.config.addCustomCategoryComment(CATEGORY_VERSION, "Used for Configuration Version Control. DO NOT CHANGE THIS.");
-            config.saveConfig();
+            VersionConfig config1 = createNewVersionConfig(config.configFile, config.targetVersion, config.modid);
+            return config1;
         }
 
         return config;
